# openapi3.yaml
openapi: 3.0.1
info:
  title: heartbeat manager
  description: heartbeat storage service
  version: 1.4.1
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
  /heartbeat/expired/{heartbeatDuration}:
    parameters:
      - $ref: '#/components/parameters/heartbeatDuration'
    get:
      operationId: GetExpiredHeartbeats
      tags:
        - Heartbeat
      summary: >-
        gets ids of tasks with expired heartbeats based on the specified
        duration
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/idList'
...
components:
  parameters:
    heartbeatDuration:
      in: path
      name: heartbeatDuration
      description: specific amount of milliseconds to be considered as expired heartbeat
      required: true
      schema:
        type: integer
...
  schemas:
    idList:
      type: array
      items:
        type: string
      example:
        - id1
        - id2
...